output-service/requirements.txt
nats-py==2.3.1
redis==5.0.1
python-telegram-bot==20.6